+++ app/controllers/steps/income/client/employment_details_controller.rb
@@ -2,26 +2,14 @@ module Steps
   module Income
     module Client
       class EmploymentDetailsController < Steps::IncomeStepController
-        def edit
-          @form_object = EmploymentDetailsForm.build(
-            employment_record, crime_application: current_crime_application
-          )
+        include Steps::Income::Client::EmploymentUpdateStep
+
+        def advance_as
+          :client_employment_details
         end
 
-        def update
-          update_and_advance(EmploymentDetailsForm, record: employment_record, as: :client_employment_details)
-        end
-
-        private
-
-        def employment_record
-          @employment_record ||= employments.find(params[:employment_id])
-        rescue ActiveRecord::RecordNotFound
-          raise Errors::EmploymentNotFound
-        end
-
-        def employments
-          @employments ||= current_crime_application.employments
+        def form_name
+          EmploymentDetailsForm
         end
       end
     end
